requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.2.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.110.1
ffmpeg-python==0.2.0
flower==2.0.1
future==1.0.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
humanize==4.9.0
idna==3.7
jose==1.0.0
kombu==5.3.7
passlib==1.7.4
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.1
pyasn1==0.6.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==4.5.4
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.37.2
tornado==6.4
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
##uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
locust==2.25.0
google-cloud-storage==2.15.0
google-cloud-core==2.3.2